+++ cameras/pinholeCamera.h
@@ -6,13 +6,13 @@
 #include "../ray.h"
 class PinholeCamera : public Camera {
   public:
-    double d;
+    float d;
 
     PinholeCamera(const Vec3& _camPos, const Vec3& _camForward, const std::shared_ptr<Film>& _film) : Camera(_camPos, _camForward, _film) {
       d = 1;
     };
 
-    bool getRay(double u, double v, Sampler& sampler, Ray& ray) const {
+    bool getRay(float u, float v, Sampler& sampler, Ray& ray) const {
       u = -u;
       v = -v;
       ray = Ray(camPos, normalize(d*camForward + u*camRight + v*camUp));
