+++ cameras/thinLensCamera.h
@@ -4,24 +4,24 @@
 #include "../camera.h"
 class ThinLensCamera : public Camera {
   public:
-    double a;
-    double b;
-    double f;
-    double Fnumber;
+    float a;
+    float b;
+    float f;
+    float Fnumber;
     Vec3 lensCenter;
-    double lensRadius;
+    float lensRadius;
 
-    ThinLensCamera(const Vec3& _camPos, const Vec3& _camForward, const std::shared_ptr<Film>& _film, const Vec3& focusPoint, double _a, double _Fnumber) : Camera(_camPos, _camForward, _film), a(_a), Fnumber(_Fnumber) {
+    ThinLensCamera(const Vec3& _camPos, const Vec3& _camForward, const std::shared_ptr<Film>& _film, const Vec3& focusPoint, float _a, float _Fnumber) : Camera(_camPos, _camForward, _film), a(_a), Fnumber(_Fnumber) {
       b = (focusPoint - this->camPos).length() - a;
       f = 1/(1/a + 1/b);
       lensRadius = 0.5 * f/Fnumber;
       lensCenter = this->camPos + a*this->camForward;
     };
 
-    bool getRay(double u, double v, Sampler& sampler, Ray& ray) const {
+    bool getRay(float u, float v, Sampler& sampler, Ray& ray) const {
       Vec3 sensorPos = this->camPos + u*this->camRight + v*this->camUp;
       Vec3 sensorToLensCenter = normalize(lensCenter - sensorPos);
-      double cos = dot(this->camForward, sensorToLensCenter);
+      float cos = dot(this->camForward, sensorToLensCenter);
       Vec3 objectPos = sensorPos + (a + b)/cos * sensorToLensCenter;
 
       Vec2 l = sampleDisk(sampler.getNext2D());
