+++ lights/pointLight.h
@@ -13,9 +13,9 @@ class PointLight : public Light {
       return this->power;
     };
 
-    RGB sample(const Hit& res, Sampler& sampler, Vec3& wi, Vec3& samplePos, double& pdf) const {
+    RGB sample(const Hit& res, Sampler& sampler, Vec3& wi, Vec3& samplePos, float& pdf) const {
       samplePos = lightPos;
-      double dist2 = (samplePos - res.hitPos).length2();
+      float dist2 = (samplePos - res.hitPos).length2();
       wi = normalize(lightPos - res.hitPos);
       pdf = 1.0 * dist2;
       return this->power;
