+++ materials/mirror.h
@@ -12,7 +12,7 @@ class Mirror : public Material {
       return RGB(0);
     };
 
-    RGB sample(const Vec3& wo, Sampler& sampler, Vec3& wi, double& pdf) const {
+    RGB sample(const Vec3& wo, Sampler& sampler, Vec3& wi, float& pdf) const {
       pdf = 1.0;
       wi = reflect(wo, Vec3(0, 1, 0));
       return 1.0/absCosTheta(wi)*reflectance;
