+++ shape.h
@@ -7,6 +7,6 @@
 class Shape {
   public:
     virtual bool intersect(const Ray& ray, Hit& res) const = 0;
-    virtual Vec3 sample(Sampler& sampler, Vec3& normal, double& pdf) const = 0;
+    virtual Vec3 sample(Sampler& sampler, Vec3& normal, float& pdf) const = 0;
 };
 #endif
