+++ skys/ibl.h
@@ -12,11 +12,11 @@ class IBL : public Sky {
     int width;
     int height;
     float* data;
-    double intensity;
-    double offsetX;
-    double offsetY;
+    float intensity;
+    float offsetX;
+    float offsetY;
 
-    IBL(const std::string& filename, double _intensity, double _offsetX, double _offsetY) : intensity(_intensity), offsetX(_offsetX), offsetY(_offsetY) {
+    IBL(const std::string& filename, float _intensity, float _offsetX, float _offsetY) : intensity(_intensity), offsetX(_offsetX), offsetY(_offsetY) {
       int n;
       data = stbi_loadf(filename.c_str(), &width, &height, &n, 0);
     };
@@ -25,12 +25,12 @@ class IBL : public Sky {
     };
 
     RGB getColor(const Ray& ray) const {
-      double phi = std::atan2(ray.direction.z, ray.direction.x);
+      float phi = std::atan2(ray.direction.z, ray.direction.x);
       if(phi < 0) phi += 2*M_PI;
-      double theta = std::acos(ray.direction.y);
+      float theta = std::acos(ray.direction.y);
       
-      double u = std::fmod(phi + offsetX, 2*M_PI)/(2*M_PI);
-      double v = std::fmod(theta + offsetY, M_PI)/M_PI;
+      float u = std::fmod(phi + offsetX, 2*M_PI)/(2*M_PI);
+      float v = std::fmod(theta + offsetY, M_PI)/M_PI;
       
       int w = (int)(u * width);
       int h = (int)(v * height);
